major refactor of components
