+++ src/Blocks/BlocksExample.php
@@ -55,6 +55,9 @@ class BlocksExample extends AbstractBlocks
 
 		// Filter block content.
 		\add_filter('render_block_data', [$this, 'filterBlocksContent'], 10, 2);
+
+		// Output inline css variables.
+		\add_action('wp_footer', [$this, 'outputCssVariablesInline']);
 	}
 
 	/**
